repository: Add empty vote in ApplyGitattributes
In https://gitlab.com/gitlab-org/gitaly/-/merge_requests/6573 we made
ApplyGitattributes as a no-op. However, removing the voting logic is
causing https://gitlab.com/gitlab-org/gitaly/-/issues/6315. This is
because if praefect doesn't receive votes and it considers
all secondaries are outdated and in need of replication.

To fix it, we add an empty logic back.

Changelog: fixed