ci: Fix permission denial issue if cache is missing
In the prior version, _build and artifacts such as binaries belong to
the 9999 user ID. However, all files and directories inside GOCACHE
and GOMODCACHE belong to root user. In test:* jobs, the test target run
unprivileged. They still have access to such files, but don't have
write access. So, if go mod artifacts are not available (cache misses
for example), test jobs won't be able to fulfill them due to
permission denial.

That's not a problem if the cache is re-used so that test jobs don't
need to re-run go mod *. It's a bit odd because the cache is supposed
to speed up tests, not to correct the permission issue.

In this version, we create the _build directory, run the
`go mod download` then chown to change the owner of the whole directory
tree to 9999. This fixes the mentioned permission issue. Now the tests
run successfully without depending on cache availability.