ci: Add Go 1.23 to the test matrix
Go 1.23 was released (https://tip.golang.org/doc/go1.23). This version
adds many nice perks, which are useful to Gitaly.

The testing version was added to the build images:
https://gitlab.com/gitlab-org/gitlab-build-images/-/merge_requests/844

It's worth noted that the minimum Git version of the image was bumped
to 2.45. This version aligns with our minimum supported Git version.
https://gitlab.com/gitlab-org/gitlab-build-images/-/merge_requests/835