gitlab-ci: Change git version for reftables
The reftable tests were running on the 'master' branch of Git and were
overriding the Git version to be "v99.99.99". This was because there
were a few patches we relied upon which was in master but not yet in a
released version.

Since then Git v2.46.0 has been tagged, so let's change the required
version to be the same. This also fixes the breaking reftable tests,
since this was caused due to the auto-compaction fixes for reftables
which enable compaction for non-locked tables, which also landed in Git
master.