git: Change the version requirement for symref-update
The support for symref update was aligned to Git version "v99.99.99"
because the patches were still in master and not in a tagged release.
Now that Git 2.46.0 has been released with the symref update patches
included, let's modify the requirement.

Also fix `gittest.IfSymrefUpdateSupported` to check if the featureflag
is enabled.