gitlab-ci: Remove 'allow_failure' for reftable tests
The reftable tests were allowed to fail before, since the Git patches
they depended on weren't still merged to master and there was no tagged
release with them. Now that Git v2.46.0 is tagged and released, it
contains all the required changes. Also we have now stabilized the
reftable changes, let's remove the 'allow_failure' in this context.