fix: Address non-constant format string errors
The updated version of golangci-lint appears to be more aggressive in
checking that the first argument to `printf` style calls are constant
strings. This impacts our various `structerr` error constructors and
some other miscellaneous callsites.

Appease the linter by ensuring the first argument is a constant string.
In the case of `structerr` constructors with only a single argument, use
"%s" as the format string.

# Conflicts:
#	internal/gitaly/service/diff/diff_blobs_test.go