diff: Test `DiffBlobs` request with new empty file
When the `DiffBlobs` RPC receives a request blob pair with a NULL OID,
Gitaly substitutes it with an empty blob ID. This is done to allow blob
addition/deletion diff generation because git-diff(1) does not support
NULL OIDs. If the added/deleted blob is also empty, the RPC returns an
error because the pre-image and post-image blobs being diffed are
identical. This results in git-diff(1) generating no output at
consequently the diff parser fails.

This is a bug and will be fixed in a future commit. Add a test case
demonstrating the failure and expected error.