git: Set files ref-backend for git-clone(1) and git-init(1)
When we initialize a repository via git-init(1), or clone a repository
via git-clone(1), we don't specify the reference backend to use. This is
OK, since Git, defaults to the files backend.

We eventually want to switch to the reftable backend. So let's add code
to make the reference backend explicit by specifying the `--ref-format`
flag.

In the upcoming commit, we'll add a feature flag to allow us to swap
out the backend with the reftable backend.