git: Change ref-format behind a featureflag
In the previous commit, we made the 'files' backend the default ref
format. Let's add a featureflag to allow us to change this to
'reftable'. This allows us to ensure that new repositories created under
Gitaly will be using the 'reftable' backend, when the flag is enabled.

For existing repositories, we'll have to use the upcoming migration
tooling in the transaction manager to perform an in-place reference
backend migration.

Also set the flag to be enabled, if the `GIT_DEFAULT_REF_FORMAT` is set.
Because the `--ref-format` flag would take precedence otherwise and
there would be mismatch in the tests expectation.