storagemgr: Correctly handle default branch updates
Earlier, default branch updates where manually handled, wherein we would
lock the '$GITDIR/HEAD' file and update its contents. This doesn't work
with the 'reftable' backend. So we introduced the 'symref-update'
functionality in Git to allow us to update default branches via
git-update-ref(1). This provides transaction support needed and we can
use this to propagate the default branch updates via the
reference-transaction hook, which allows us to vote on such updates.

Unfortunately, the implementation in the storagemgr is linked to
reftables. But it shouldn't be so, since the feature can also be used
with the files backend. Fix the implementation to be more generic.