storagemgr: Accept `context` for `TransactionManager`
The `NewTransactionManager` doesn't accept a context and instantiates
one from scratch. This is okay, but in tests we use the context to
propagate flag values.

Let's modify the function to accept a context. This ensures that the
flags set in tests are correctly propagated. This shouldn't affect the
normal flow, since in the normal flow this is called by
`NewPartitionManager` which also receives a context and both are not
tied to the request <> response cycle.