Makefile: Use Go build ID to derive GNU build ID
We have some hacks in our Makefile that allow for the generation of
reproducible GNU build IDs. The way this is implemented is kind of ugly:

  1. Build an intermediate binary with a fixed GNU build ID.

  2. Extract the Go build ID and hash it to derive a deterministic GNU
     build ID.

  3. Copy the intermediate binary into place, replacing the fixed GNU
     build ID from step (1) with the computed one.

This is rather complex and also wasteful, as we essentially have to copy
hundreds of megabytes a second time whenever we build our binaries.

To address this, we have upstreamed 089cc68676 (cmd/link: allow deriving
GNU build ID from Go build ID ID, 2023-09-16) that does this for us in
the Go build chain. When passing the `-B gobuildid` linker flag, then
the linker will compute the GNU build ID from the Go build ID and thus
do essentially the same as we did manually in Gitaly. This new feature
has been released with Go v1.22.

Now that we have bumped the minimum Go version in 60d22fac1 (go: Bump
Go's minimum version to v1.22, 2024-08-23) we can finally use this new
flag. This allows us to simplify our build instructions. And as a nice
benefit, it also improves build times by a couple of seconds on my
machine, which is especially noticeable for incremental builds:

    Benchmark 1: make (revision = HEAD~)
      Time (mean ± σ):      5.197 s ±  0.060 s    [User: 11.138 s, System: 5.807 s]
      Range (min … max):    5.117 s …  5.318 s    10 runs

    Benchmark 2: make (revision = HEAD)
      Time (mean ± σ):      2.009 s ±  0.046 s    [User: 7.493 s, System: 3.011 s]
      Range (min … max):    1.925 s …  2.075 s    10 runs

    Summary
      make (revision = HEAD) ran
        2.59 ± 0.07 times faster than make (revision = HEAD~)