Define Transaction interface in storage package
There was previously a cyclic dependency between the 'git' and 'storage'
packages. We've solved this now by splitting the implementation related
bits out from the 'git' package and only type definitions and interfaces
there. With the cycle out of the way, we can now define the Transaction
interface in a single location and reuse it across all the sites that
previously were using one of the other interface definitions.

This commit defines the interface of a Transaction in the storage
package. We also replace duplicate definitions of the interface in
other packages.