Implement a node manager
Each Gitaly node may have zero or more storages configured. Currently
routing transactions to correct storages is handled by PartitionManager.
PartitionManager has grown to be too complex and is implementing
responsibilites of multiple layers. We've since split out the management
of a single storage into the StorageManager type.

This commit goes a step further and implements a replacement for the
remainder of PartitionManager. The new manager type is the top-level
type through which storage on Gitaly will be accessed. It sets up the
storages, and provides access to them based on their name.

This commit just implements the new type, and we'll wire it into Gitaly
in a later commit.