Include all quarantined objects in the committed packfile
TransactionManager is currently dropping unreachable objects from
the transaction. This was done previously as we didn't have a good
way to figure out whether the repository contained all of the
dependencies of the unreachable objects. Now that we can do so with
the 'rev-list --missing' walk, there's little reason to drop the
unreachable objects anymore.

This commit changes transaction manager to commit all objects
introduced during the transcation, even the unreachable ones. We do
so by walking all of the quarantined objects, and figuring out their
dependencies. This ensures that even the unreachable objects that
are being committed have all of their dependencies satisfied.

This leads to a simpler interface, as the user of the transacstion
doesn't explicitly have to note which unreachable objects to include
anymore. They can simply write objects into the quarantine. If some
objects should not be included, the user would simply not write them
there.

In addition, this opens the door for future optimization by avoiding
the packfile rewriting. If the transaction's objects are already
packed in a single packfile, we could techincally just log that
directly as we now verify all of the objects in the packfile would
have their dependencies met.