Take all transaction's parameters as a struct
TransactionManager is currently taking parameters in a mixed manner
with some coming as independent parameters and some coming via an
options struct. This this is inconsistent and leads to an awkward
and confusing interface. Change TransactionManager to take in all
of the options in via a struct.

The new options type switches around the read-only flag to a write flag.
This works out better as a default since we'd want all transactions
that can be read-only be read-only which is now the default. If someone
attempts to accidentally write in a read-only transaction, and error
will be raised and the issue can be fixed. This is not the case if
someone only performs reads in a write transaction, so it's easier to
miss write transactions that don't perform any writes.

We define the new options type in the storage package in anticipation
of defining the interface of a partition there. BeginOptions will be a
parameter in the interface's Begin method.