Move DBManager to its own package
The keyvalue package contains the KV interface of Gitaly, and some
generic wrappers around it. We'd like to expose a transaction
interface from the storage package. The DBManager however causes
a cyclic dependency as it relies on the config package where as the
rest of the keyvalue package doesn't. Move the DBManager to its own
package to break the cycle.