Merge storagectx package into storage
storagectx package was a workaround that was needed in the past as
the 'git' package would have had cyclic depencies with the 'storage'
pacakge if we define the Transaction interface there. We've since
restructured the 'git' package to solve the structural issue there
and moved the Transaction interface into the 'storage' package.
There's no reason for the storagectx package to exist anymore. This
commit merges the package into the 'storage' package to simplify.