storagemgr: Only manually track 'HEAD' ref for files-backend
To update the default branch, there are two flows which can take place.

When symref-update is disabled (only for files-backend), we manually
call `repo.setDefaultBranchManually` to update the default branch. There
we manually write to the 'HEAD' ref and call
`tx.MarkDefaultBranchUpdated()` so that the transaction records the
'HEAD' ref changes.

When symref-update is enabled (for files, reftable backend), we call
`repo.setDefaultBranchWithUpdateRef` to update the default branch. Here,
we use 'git-update-ref(1)' to actually do the update. This means the
update is caught by the reference-transaction hook. The hook then adds
the updates to the transaction manager as reference updates.

In the transaction manager, for recording reference updates, there
are two flows, based on the reference backend present.

For the reftable backend, we apply the reference updates via
'git-update-ref(1)' and track the changes in list of reftables present.
Since root-refs like 'HEAD' are also present in the same table, we don't
have to do anything special.

For the files backend, we apply the references via 'git-update-ref(1)'
and track the changes in the '$GIT_DIR/refs' and '$GIT_DIR/packed-refs'.
This means root-refs like 'HEAD' are not tracked here. Instead changes
to the 'HEAD' ref are tracked when the `repo.setDefaultBranchManually`
is set.

So with this context, we need to ensure that the files backend also
works when using symref-update. So we need to track default branch
updates in the reference transaction hook and call
`repo.setDefaultBranchManually` correspondingly. The transaction manager
however records the 'HEAD' file update unconditionally even for the
reftable backend. This doesn't make sense the 'HEAD' ref is a dummy in
the reftable backend. So let's make this conditional.

Note: Another possible solution for this is to actually bake in tracking
of 'HEAD' ref in `verifyReferencesWithGit` similar to how the
'packed-refs' file is tracked. But that change would be little more
involved. So this seems like a simpler fix for now. Eventually when the
symref-update becomes default, we could probably strip away a lot of
this code and make it much simpler.