storagemgr: Throw error if symref-update is not supported
In `applyReferenceTransaction` we apply the reference updates as part of
applying the logged transaction. Since the transaction manager isn't
linked to the request <> response cycle, it doesn't get the context
filled with featureflag values.

This means while we're rolling out the newer Git versions (v2.45.0 to
v2.46.0) or rolling out 'symref-update' via featureflags, those flag
values are not seen by the transaction manager. So we need to be more
defensive in support all the features while also erroring out for
unsupported features. One such situation is that if the Git version is
v2.45.0 and there is a symref update, we know we cannot perform this
update since a minimum of Git v2.46.0 is required.

Currently, while we check for symref-support we don't error out. This
causes unwanted behavior, where a symref update for the 'HEAD' ref is
then applied as a regular ref update for 'HEAD', wherein it gets
deleted, since the 'NewOID' value wasn't set. Let's be more careful here
and error out as necessary.