storagemgr: Skip testing with symref-update for old Git versions
For running tests with 'symref-update', we want to ensure that the Git
version is v2.46.0 and above. So let's a check before running the tests
and skip as necessary. This will go in hand with the changes we plan to
introduce where we tighten the checks in the transaction manager around
the symref-update feature usage.