Rename references of TransactionManager to Partition
We're going to split the TransactionManager into its own package.
PartitionManager currently depends on a private interface of
that TransactionManager fulfills. We need to make this public as
we'd like to export the factory interface it's using currently to
create TransactionManagers. This will help us avoid drilling
TransactionManager's dependencies through the PartitionManager
unnecessarily. If we were to just export this, the name would
conflict with the existing TransactionManager type. This commit
renames the references to TransactionManager within PartitionManager
to Partition. This allows us to export the interface from the
package without a naming conflict, and better matches what it
logically represents. TransactionManager is more of an implementation
detail of a partition.