Don't rely on TransactionManager internals in PartitionManager tests
PartitionManager has two tests that are relying on the internal
behavior of TransactionManager. These tests are aiming to assert
that correct context is passed down, and the cancelation in the
context is respected. As we're about to move TransactionManager
into its own package, it would be better not to rely on these
details. Use a mock to exercise these cases instead.