Use raw KV store in TransactionManager tests
TransactionManager is currently using a prefixed KV store in its
tests. While this is what it would actually be using when all
components are wired up, its an implementation detail of the
partitioning layer. TransactionManager itself shouldn't care about
what kind of a KV store it gets as long as it gets one. Stop using
this in the tests as it's an irrelevant detail at this layer.