Export PartitionFactory type
PartitionManager is coupled tighter than necessary to TransactionManager
as it is responsible for instantiating a TransactionManager. To break this
coupling, we promote PartitionFactory to a proper type. The factory knows
how to instantiate a TransactionManager which allows the PartitionManager to
be oblivious to this. This also allows us to move some of the
dependencies specific to the TransactionManager to the factory, and later
stop drilling them through PartitionManager.

While we implement and export the type here, we'll dependency inject it in
a later commit.