Remove StorageKV method
StorageKV is a method implemented on PartitionManager for specifically
performing key-value operations on a transaction. The method doesn't
really belong on the type.

1. Instead of a KV specific method, we should just start a transaction
   and do the key-value operations there.
2. We're also testing the written key-value state in tests. This is
   not the responsibility of PartitionManager. TransactionManager is
   responsible for the key-value operations.

Remove the method and inline it in the only calling location in the
Raft package. We add a KVOnly option on the transaction to be able
to start a transaction without any repositories in the snapshot.
The Raft package is the only user of this functionality for now.
However, the Raft package is layered incorrectly. Raft should live
under partition, not perform writes in it through transactions. As
this is rather temporary, and the integration needs to be rewritten,
we don't add tests for the new option for now. Instead, we'll remove
it soon enough through the on-going refactoring work.