Move log consumer interface to storage package
Log consumer interfaces will be for now needed by both PartitionManager
and TransactionManager. Move them to storage package where they can
be imported by both.

Move the consumer factory to be colocated with PartitionManager as it's
only used by it. In reality, PartitionManager shouldn't need the consumer
factory, and it should be possible to handle the log management entirely
within TransactionManager. There's a dependency between the types for
now though. Let's leave this as is, and solve it later.