backup: Validate repository before calling updateref
In `backup::localRepository.ResetRefs` we try to read the references
from a local repository. There are situations where the repository
doesn't even exist. In this case, we detect the failure when we try to
read the 'ObjectHash' of the repository.

This isn't ideal, since the value of 'ObjectHash' is cached and later
when we do end up creating a repository, the 'ObjectHash' would still
return the same error we get now. So instead, let's do the right thing
and validate the repository at the start.