Split TransactionManager into its own package
The storagemgr package has grown quite unwieldy. Worse than the size,
the types in it have grown unncessary dependencies to each other.
We'd like to define a clear interface between all of the types, with
separate layers for:

1. Managing node level concerns like routing to correct storage.
2. Managing storage level concerns like managing the key-value store
   and routing to correct partitions.
3. Managing partition level concerns such as transactions.

This commit takes the first step towards that by splitting the
TransactionManager into `partition` package which will host the
partition layer and its concerns.

The commit only moves code into the new package, and fixes import paths.

While there are plenty of clean ups that could be done including the
type names in the new package, we'll leave those for later to keep this
change reasonable size.