diff: Fix diff generation for empty files
The `DiffBlobs` RPC accepts pairs of blob revisions used to generate a
batch of diffs. Normally these blob pairs should should not contain a
set of revisions that resolve to the same OID because git-diff(1) does
not generate patch or raw formatted output is this scenario. From the
client perspective, it also doesn't make sense to make these types of
requests because they should already know that there is no difference
between identical blobs.

The `DiffBlobs` RPC also accepts a NULL OID as either the left of right
blob revision in order to generate addition or deletion diffs. Normally
git-diff(1) does not understand these revisions, but `DiffBlobs` works
around this by rewritting the NULL OID to the OID of an empty blob and
temporarily creating the object in the repository.

Unfortunately, this makes it possible for the resolved blob IDs to match
and cause the diff parser to error out. This happens when one side of
the diff uses a NULL OID and the other side any revision that resolves
to an empty blob. This would be seen for diff requests where an empty
file is either being added or deleted.

Fix the error by detecting when the rewritten blob IDs match and
returning early with an empty diff. This works because identical blobs
would not produce patch output anyways.