Move fields used only by storageMgr there from PartitionManager
We're going to split PartitionManager into two types, one which handles
node level concerns such as routing to correct storages, and one which
handles storage level concerns such as routing to correct partition. To
do so, we need to refactor a bit.

PartitionManager contains some fields it doesn't use itself but only
holds so they can be used by the storageMgr. Move these fields to the
storageMgr where they are actually used.