Split Begin
Begin is currently handling node and storage level concerns.
It first routes to the correct storage, and then runs the logic to
start a transaction against a partition.

Split the method in these two parts so we can split PartitionManager
into two different types that will handle node and storage level
concerns.