Scope PartitionManager tests to storageManager
We're going to replace PartitionManager with a new type that focuses
just on node level concerns such as routing to correct storage. Doing
so, we'll promote storageManager to an exported type later as it
implements the actual storage related concerns. To prepare for this
split, rescope the PartitionManager tests to storageManager. Some tests
that exercise multiple storages are removed as a given storage manager
only manages a single storage.