Take unscoped metrics in newStorageManager
newStorageManager takes in the scoped metrics via its constructor.
Change it to take the uncsoped metrics for now so we don't have
to export both types. The unscoped metrics are in fact entirely
metrics of storage manager, PartitionManager just happens to be the
one collecting them. We'll later handle the collection separately
in Gitaly's main.