Export TransactionOptions from storage package
TransactionOptions are the parameters Begin takes. As we need to
export an interface for now to call Begin, we need to export the
options type from storage.

TransactionOptions duplicate some of the options in the BeginOptions
struct. This is an existing wart of the implementation and will go
eventually as we progress with the refactoring. Exporting the type
here for now allows us to split PartitionManager and StorageManager
into separate packages.