Remove IsClosed from Partition and TransactionManager
StorageManager's tests are asserting internal state of the manager
in its tests. Before we make assertions, we have to wait for all
in-flight changes to finish so the state stabilizes to avoid flakes.
`IsClosing()` is used only to do this in tests. It returns true if
the `Close()` method on the `Partition` was called. If the partition
is closing, it will soon be removed from the active partitions
map which could lead to flakes if we don't wait for it before asserting
the state. As this is purely a test related synchronization concern,
let's handle this locally in the tests.

This commit removes the `IsClosing()` method and simply stores on the
mock used in the test whether `Close()` has been called. If so, we know
we need to wait for the partition to exit first.