Run entire OptimizeRepository in transaction
OptimizeRepository is currently not running all of its logic in a
transaction. Expand the maybeStartTransaction block to do so except
for the cfg parsing which still needs to be outside given it is needed
to call the function.

This exposes a bug where the method is not accounting for relative path
rewriting, and is constructing repository locks using with relative paths
containing the snapshot prefix. This allows multiple OptimizeRepository
calls to run concurrently on a repository.