Prevent concurrent housekeeping with transactions
OptimizeRepository is acquiring a repository specific lock before
running to prevent concurrent housekeeping operations from targeting
the same repository. The lock key is constructed using the relative
path of the repository.

If transaction are used, transactions using different snapshots will
have a different relative path for the same repository as the relative
path contains the snapshot prefix. This may lead to concurrent
houskeeping operations on the repository.

Fix this issue by constructing the key using the original, non-rewritten
relative path of the repository.