Always start a transaction in OptimizeRepository
OptimizeRepository is called from multiple different contexts:
- As part of the daily maintenance. This does not set up a transaction
  and relies on OptimizeRepository to do so.
- As part of FetchIntoObjectPool. While the RPC runs in a transaction,
  it's committed by the time OptimizeRepository is called. This thus
  relies on the transaction to be started in OptimizeRepository.
- Through the OptimizeRepository RPC runs in a transaction. The RPC's
  transaction is used directly by the housekeeping manager.

This is a bit complicated, especially as we're soon about to use multiple
transactions in OptimizeRepository for reference and object repacking.
Simplify by always removing the option to reuse the existing transaction
in the context and instead always start a new one. While this does lead
to a duplicate transaction when called in the context of the
OptimizeRepository RPC. This is a minor issue though, and we can improve
it later.