Pack references and objects in different transactions
Reference and object repacking currently run in the same transaction.
This is not optimal as this increases the chance of conflicts.

Packing reference conflicts with reference deletions. The longer the
transaction runs, the higher the chance there is that a reference
deletion is concurrently committed, and thus forcing the reference
packing transaction to abort. Generally reference packing is fairly
fast, and deletions not that common. However, we're currently repacking
objects in the same transaction which can be significantly slower. This
delays committing the reference packing which leads to a high chance of
conflicts.

Run the two operations in different transactions to reduce the likelyhood
of reference repacking encountering conflicts. As the parent transaction
used in OptimizeRepository is no longer performing the changes itself,
it's made read-only.