Abort transactions with concurrent changes to packed-refs file
TransactionManager is currently replaying the reference changes made
in a transaction to verify they apply against the repository's current
state. This also resolves any potential conflicts in the packed-refs
file. If two transactions delete references from the packed-refs file
concurrently, they'd produce different/conflicting packed-refs files.
As the reference verification replays the changes, the latter transaction
is effectively rebased on top of the earlier transaction and the conflict
resolved by committing the packed-refs file that was produced during the
reference verification.

While resolving the conflicts reduces transaction aborts due to conflicting
changes, it causes performance issues. The packed-refs file can be large.
In gitlab-org/gitlab it's currently about 2GB. Deleting a reference from
the packed-refs file requires a complete rewrite of the file. Currently
that operation takes about eight seconds. During these eight seconds, all
other writes would be stuck in the commit queue waiting for their turn.
This led to errors and latency increases as we attempted to deploy
transactions on production canary. To avoid the performance issues, we'll
instead reject transaction that attempt to perform conflicting changes in
the packed-refs file. Trading increased aborts for better performance is
a good choice as transaction processing being blocked affects all writes,
where as the aborts only affect reference deletion and repacking.

This commit stops rewriting the packed-refs file when a transaction is
being committed. We do so by taking the final packed-refs file from the
transaction's snapshot and committing that directly. When replaying the
reference changes, we replace the repository's current packed-refs file
with the one from the transaction. Any possible reference deletions that
we'd replay would no longer need to modify the packed-refs file. The
reference would already have been deleted from there during the
transaction.

As simply replacing the file could lead to lost updates, we also add
conflict checks that abort transactions that concurrently change the
packed-refs file. This is done by recording the inode of the packed-refs
file as it was when the transaction began, and comparing it to the inode
of the packed-refs file in the target repository at commit time. If they
differ, there's been a concurrent modification of the file and we abort
the transaction.

As a special case, reference packing operation conflicts with reference
deletions even if they didn't lead to packed-refs changes. This is because
the reference packing would move the to-be-deleted loose references to the
packed-refs file. Solving this would require rewriting the pack in the
critical section so this leads to aborts.