Restructure TransactionManager metrics
We currently don't have a place to put the top-level metric vectors
that each TranscationManager should be recording metrics to. This
commit introduces a Metrics struct that contains that global metrics
that can then be scoped to a specific TransactionManager. This way we
can place all of the metric definitions there and pass them down to
the TransactionManagers fully scoped and ready to use.