ci: Increase Postgres max_connections and cache
While working on https://gitlab.com/gitlab-org/gitaly/-/issues/6329, I
noticed that some Praefect tests were consistently failing due to the
connection limit being exceeded. This was likely caused by more
concurrently-running tests due to more resources being available on the
new runner VMs.

As we don't maintain a DB connection pool in tests, we need to increase
the maximum allowed client connections. While we're here, we should
increase the amount of cache available for writes too (the default is
128MB according to
https://www.postgresql.org/docs/14/runtime-config-resource.html)

Note that we don't increase max_connections for the Postgres instance in
the pgbouncer test, because pgbouncer is responsible for maintaining a
connection pool.