generic: Swap out `CommandFactory` with `RepositoryExecutor`
In a lot of places in our code base, we create a `CommandFactory` and
use `gitCmdFactory.New`. Instead let's switch to using
`RepositoryExecutor::Exec` which does the same thing. This commit is
mostly a no-op where we only change the function used to ensure that we
cleanup the codebase to mostly use the `RepositoryExecutor` everywhere.

This allows us to make it more consistent and eventually also focus on
propagating the `RepositoryExecutor` through the layers.