gitcmd: Remove `symlinkPointsToGitDir` configuration
In Git versions 2.45.1 and 2.45.2, the `symlinkPointsToGitDir` fsck
exists and warns if a repository contains symlinks pointing inside the
Git directory. Gitaly sets `transfer.fsckObjects=true`, which
consequently upgrades these warnings to errors and can break existing
repositories. For this reason, Gitaly explicitly ignores these checks.

In subsequent Git versions, this configuration is reverted and removed.
To avoid and unknown configuration warning, Gitaly only adds the
configuration for the required versions. Now that the minimum Git
version is 2.46.0, remove this workaround.