Move partition reference counting to partition from transaction
Partition lifecycle is managed by tracking how many transactions are
still using it. We want to soon return partition handles from the
storage manager as starting a transaction belongs to the partition
layer rather. Move the reference counting implementation from the
finalizableTransaction type to partitionHandle that fits better
for this purpose.