Remove KVOnly option
Storage interface now exposes a method to get a partition. Using that,
callers now have direct access to start a transaction on a partition
and thus also its options. There's no need for the KVOnly option
anymore as the use case can be achieved by directly starting a
transaction against the partition without going through the Begin
wrapper on Storage which will eventually be removed.