Start a transaction directly against the partition
Transaction middleware knows how to start a transaction against
specified partition. It currently does so by using the `Begin`
method on a Storage. We'd like to remove the method as it violates
the layering. Starting a transaction belongs to the partition layer.

This commit starts a transaction for a partition scoped RPC directly
against the specified partition. This will allow us to remove the
option to specify a partition ID in Begin where it does not belong.
This also makes it significantly easier to see what is going as the
partition and repository scoped handling are no longer intertwined.