Remove partition ID from Begin
PartitionID was added at some point on Begin to enable accessing
a partition by its ID rather than a relative path. Now that we have
a better interface to do so with GetPartition, remove the option from
Begin and update all of the call sites. This allows us to simplify the
method and will make it easier to also remove it in the future.