Don't inject Node when transactions are disabled
In a recent refactoring, we split the PartitionManager into two
types. While doing so, a bug was introduced where Transactions are
partially re-enabled if they've been enabled and then disabled
again in the past. This happens as we're setting up the
transacation stack if transactions have been enabled in the past
to recover any partially applied transactions before proceeding
non-transactional access. Typically we'd just set up the Node and
the recovery middleware. After the refactoring, we're injecting
the Node also to other components which may alter their behavior
as transactions have been enabled.

Transactions have been previously enabled on production canary and
this issue would affect that environment.